ui: Avoid the use of "width-request" for the search frame
This commit changes the search_entry and replace_entry text fields to
use width-chars instead of width-request to determine their size, and
changes the other buttons in the search frame to fill the space allotted
to them rather than hard-coding pixel values. This means that the toggle
buttons change size when the "Replace" and "Replace All" buttons are
visible, so they line up as they do in the mockups.

Avoiding "width-request" properties on widgets makes Builder more
flexible for use on different screen sizes and with different languages.

https://bugzilla.gnome.org/show_bug.cgi?id=768777