editor-frame: Add squiggly underlines for invalid regex
When an invalid regex is entered in the search entry field, it's
difficult for the user to tell if there's an issue with their input or
if the search genuinely returned no results. This commit addresses that
ambiguity by adding a wavy underline when an invalid regular expression
is entered in the search field, or an invalid replacement string is
entered in the replace field (which can contain backreferences). The
color white is used so that it can still be seen when the background
is red (which happens when a search returns no results).

https://bugzilla.gnome.org/show_bug.cgi?id=768812