editor-frame: Move the logic for enabling/disabling replace actions
There are many different conditions that determine whether Replace or
Replace All operations would make sense, so similar logic needs to be
run in many callbacks to try to ensure the buttons' sensitivity reflects
the state of the search. This commit centralizes that logic into
update_replace_actions_sensitivity() which makes the code more
maintainable.

https://bugzilla.gnome.org/show_bug.cgi?id=768812