color-picker: add palette menu and popover
We now use the same menu and popover as for the
color widgets:

menu with rename and remove.
rename popover for the remove choice.

The old flat GtkEntry used for the name
and to edit it is removed.

The contextual menu is triggered by right click or
by the F2 key.

Please enter the commit message for your changes. Lines starting