layout-stack: add IdeLayoutStackAddin
This plugin interface can be used to extend the IdeLayoutStack. A new
plugin instance will be created for each layout stack and destroyed when
the stack is destroyed.

This can be useful to add controls to the IdeLayoutStack. Doing so instead
of using IdeLayoutView.get_controls() because it allows us to have a set
of controls once rather than multiple times in the UI, saving widget
overhead. This is more of a problem when using Popovers and Lists in that
widgetry.