editor: use single instance of tab controls per stack
Instead of creatin new popovers and lists for every document that is
opened, use a single instance of the controls per stack. This saves on
lots of popovers when lots of documents are opened as well as cutting down
on the amount of style validation that has to occur with those additional
widgets.