autotools-templates: Add a GNOME application template
This commit adds a template for GNOME applications. It's similar to the
Empty Project template but it comes with a main function that creates an
empty window, and it includes a flatpak manifest. Hopefully it will help
newcomers get started on their first project.

In the future we should probably add internationalization, gtk-doc, and
GSettings code to the template, possibly as optional components.

The flatpak manifest works when used with flatpak-builder but building
the template in a flatpak runtime from within Builder doesn't seem to
work yet.

https://bugzilla.gnome.org/show_bug.cgi?id=769564