genesis: Update the new-project action to use the new view
The New Project view used to be handled by an IdeGenesisPerspective, but
is now handled by GbpCreateProjectGenesisAddin. The application's
new-project action (which can be accessed via the menu in the GNOME
Shell top bar) stopped working when the switch was made. This commit
fixes the issue by adding an interface to the IdeGreeterPerspective to
show a genesis view, and calling it from ide-application-actions.c.

https://bugzilla.gnome.org/show_bug.cgi?id=769454