autotools: Make the Run button work when builddir != srcdir
To make the Run button work, Builder looks in a project's directory for
a Makefile, which doesn't work when the build directory differs from the
source code directory, such as when you're building in ~/.cache/.... This
commit makes Builder look for Makefiles in the build directory, as determined
by an IdeAutotoolsBuilder instance. It should be okay not to check for
Makefile.am files since this code is only used for autotools projects.

https://bugzilla.gnome.org/show_bug.cgi?id=769482