autotools-templates: Don't allow hyphens in package names
If the user chooses a project name with hyphens in it and tries to use a
variable like project-name_CFLAGS or project-name_SOURCES in their
Makefile.am, those variable names will be invalid according to Automake
and the build will fail. This fixes the issue by using the version of the
project name with underscores in the PKG_CHECK_MODULES macro.