run-manager: add concept of IdeRunHandler
This allows various plugins to hook into the run pipeline. This is useful
since we will have various ways to "run" an application. Such as through
a profiler or debugger.

We'd like the ability to simply "run again" and unify the selection of how
to run the project, which requires just a bit of an abstraction. That is
what the IdeRunHandlerInfo provides.