sysprof: add sysprof profiling plugin
This is a basic plugin for Builder that provides profiling with the sysprof
library. Sysprof has both a backend library (libsysprof-2) and a library
for UI components (libsysprof-ui-2).

This builds on those components and tries to merge the gap between
launching your project and hooking up the profiler components. It is not
yet as featureful as the standalone Sysprof UI, but simplifies the process
of launching your target project.