devhelp: drop the search provider
This has not turned out to be very useful because of our current global
search design (hard to type and jump to the result). Also, we have
ctrl+shift+f to focus the devhelp search, so it's redundant.

We might bring something back in the future that is similar once we get
a grip on how we want global search to work.