c-indenter: handle non-GNU style indentation better
While we are still waiting on our new indenter engine, we can make things
a bit better for those that are not using GNU-style C like we do in Gtk,
GLib, and Builder.

This falls back to the indent width for determining the indent after
conditions, and potential braces pre/post indent. This manages to not
effect GNU-style because it is 2 for both pre and post and should be
reflected by the current indentation.