build-commands: add prebuild and postbuild commands
This allows a configuration to specify additional commands that can be
run before or after the build system has executed its build. This might
be handy to setup various command hooks around a build system.

Additionally, these should be usable by Nil build systems that just need
to execute some commands as part of the build. (More on that later).