terminal: manage spawning pty child manually
So that we can use an alternate method of creating a TTY (such as via
the HostCommand flatpak interface), we need to take control of the spawn
process in the terminal.

We start by getting the pty master from VtePty, and then using that to
create our pts child. That FD may be passed along to flatpaks HostCommand
so that we can get a host shell with a child psuedo terminal.