terminal: learn to discover the shell from the host passwd
We need to discover the users preferred shell from the host as the shell
will not be propagated to the flatpak runtime. Which makes sense since that
shell might not be available in the runtime anyway.

Normally, vte would look at getpwent() info for the shell, but we can just
as easily shell out to the host system to discover it from /etc/passwd and
fallback to /bin/sh when that doesn't work.