subprocess: always pump default main context in main thread
If we are in the main thread (ie: the gtk thread), we want to pump the
real main loop rather than our own main context. This keeps gtk responsive
should it be necessary as well as ensures that we will get our dispatch
called (which could end up having landed on the primary main context).