egg: add EggThreeGrid
This is a GtkGrid-like container with some restraints that make it easier
to implement some designs. It only has 3 columns, and the center column
is always centered (and the left/right columns are of equal sizing).

Still to do to make this more generally useful is to nail down the
baseline support which has been tricky thus far.

This should allow us to simplify the project greeter layout quite a bit
since we no longer need all the size groups for row/column enforcement.