quick-highlight: handle delete with mark not changing position
If we delete text, there is a good chance that the mark will
not be moved, and therefore we will be out of sync. This can be
found when using the "x" command in Vim mode.

Additionally, to speed up comparison checks, this keeps a
pointer to the insert mark so we can just do a pointer
comparison in the fast path rather than a function call.