autotools: fix program discover for some GJS templates
The GJS templates that do nodist_app_SCRIPTS were not getting
discovered do to the levels of abstraction being employed. This
works around the issue by discovering where the program is
installed to $appdir. While this isn't an autotools'ism, it is
at least a GNOME'ism for now, so we should support it.