preferences: save an extra 24px in width
Currently we can't use the programming languages side-by-side
layout on screens with 1280 width. While that is less than
ideal, it could be common today with 2x rendering by the
compositor/gtk. So we need to support it to some degree.

This gets us down to 1324, so we still need to shave off another
44 pixels from the min-width.

https://bugzilla.gnome.org/show_bug.cgi?id=772700