autotools: Use the new build function for installs too
This commit fixes a bug introduced by 1c00cdd where
ide_autotools_build_task_execute_async() was called for installs,
but ide_autotools_build_task_execute_with_postbuild_finish() was called
in the callback. The mismatch caused the IdeBuildResult to stay in
"running" mode after the build finished which is reflected in the
interface as a Cancel button.