build-panel: switch to GtkTreeView for diagnostics
When building larger projects, we could potentially have a huge number of
warnings (as we've seen with Geary). When we get to the order of 2000
diagnostics, performance really goes through the floor with the the
GtkListBox design.

This switches to something that looks similar, but not identical wo what
we had previously. However, it uses a GtkTreeView and GtkListStore for
the items. To avoid using a GtkTreeModelSort, we binary search to find
the target position for the row.

Sadly, we can't get the row wrapping the same as it was for the
GtkListBoxRows. So instead we set the ellipsize on the row and use
tooltips to show the full warning information.

I'm still interested in seeing other ways to present this information,
but this gets us in a much better position of not locking up when the
situation occurs.