diagnostics: refactor IdeDiagnostician into IdeDiagnosticsManager
The IdeDiangostician was attached to the IdeBuffer directly. Instead,
IdeDiagnosticsManager is connected to the IdeContext. This is useful
because it allows us to receive diagnostics out of band from the normal
diagnose_async()/diagnose_finish() cycle.

Additionally, this allows diagnostics to be created for objects other
than the buffer in question. Such is useful with language servers where
they may process diagnostics for the project as a whole.