flatpak: fix building from within flatpaks
This ensures that we create the proper build directory in the proper
context. We need to do this from outside the flatpak (on the host) and
ensure it has enough environment to perform the task.

Additionally, this no longer users the "flatpak" binary from within the
application bundle, but instead relies on the host which should be
available given that we are 1) already getting this far, or 2) running
inside a flatpak.

This also switches to IdeSubprocessLauncher from GSubprocessLauncher.