subprocess: remove use of prctl() to set parent death signal
This cannot be used here because we will be run from a worker thread.
What this means (according to `man prtcl`) is that when the thread that
executes this subprocess (fork, execve) dies, the parent death signal will
be sent. This is not what we want as we want it only to be sent once the
entire process dies.

This can be done with prctl(), but requires that it is done from the main
thread if we want to ensure that it works as we would expect it.