configuration: add ide_configuration_get_internal_* API
This allows plugins to store additional information on an IdeConfiguration
that does not require adding new public API to the configuration object.

Think of this as a more introspection friendly g_object_set_data() API.

The data held in these "internal" values are meant more for storing things
needed at runtime, not things that are meant to be persisted to the
buildconfig file. These values are not persisted to buildconfig files.