clang: avoid completion until the translation unit is ready
We only want to complete if we have a cached translation unit for the
buffer. Otherwise, queue a request for the translation unit and ignore
this request.

If the completion request was requested by the user (ctrl+space), then we
will go through the whole process even if its a bit slow.