langserv-highlighter: Add IdeLangservHighlighter
This is a prototype of a semantic highlighter that uses the names of
symbols from the documentSymbol API call of the Language Server Protocol.
It's not ideal, because it's using fuzzy symbols and only those that
exist within the file. However, it gets us something until the Language
Server Protocol addresses semantic highlighting specifically.