terminal: fallback to ptsname() if ptsname_r() is not available
We are technically safe from race-conditions here in Builder specific
code today, as we only call ptsname_r() from the main loop. So switching
to the MT-unsafe ptsname() is fine. However, I'd like to always use the
re-entrant safe POSIX functions when possible.

This falls back to ptsname() when ptsname_r() isn't available, such as on
BSD variants.

https://bugzilla.gnome.org/show_bug.cgi?id=773880