autotools: Fix build target discovery for flatpak runtimes
ide_makecache_get_build_targets_worker() tries to find build targets by
spawning make processes in each directory with a Makefile, but since
the flatpak-build command specifies the directory it is unaffected by
the process's cwd. So finding build targets in subdirectories fails for
flatpak runtimes. This commit fixes that by using the -C option on make
to switch to the appropriate subdirectory.

https://bugzilla.gnome.org/show_bug.cgi?id=773764