clang-completion: ignore clang completion priority
The value we get from clang is not very useful for completions because it
is only valid for the initial completion set. Once the user starts typing
additional text, the priority is invalid.

As clang suggests that we cache the results and perform refiltering client
side, we might as well just use our own scoring/priority engine from the
get-go.