clang-completion: fallback to strcmp for matching priorities
It is very weird when the completion results change ordering. So we want
to keep the same ordering. For simplicity sake, just use strcmp() for
this.

However, we do want to improve the scoring engine so this is less likely
to occur.

Note that result set as we receive it from Clang are in no sorted order.
That result set order can change on every query, so stable sort on our
side is not enough to provide consistent ordering to the user.