source-view: make Tab prefer snippet movement before completion
When we are in a snippet, such as those from function parameter completion,
we would much rather prefer to use tab to move to the next snippet location
not complete the (possibly overactive) completion list.

We might need to revisit all of this as we make completion much better
about completing symbols in scope that might work, but this is better
for what we have today.

https://bugzilla.gnome.org/show_bug.cgi?id=774073