tests: add test-snippet
This is a new test that helps us start testing snippets manually. It's not
pretty, but we can start tweaking this to make sure that we handle
movements and changes properly.

The code to wait for stuff to finish is pretty touchy, since we don't have
signals to keep off in the lower layers of textbuffer/textview.