subprocess-launcher: setup IOCSCTTY on stdin_fd after setsid()
This ensures that after our forked process becomes the session leader we
setup the TTY with an ioctl() of TIOCSCTTY ensuring the terminal the
controlling terminal.

This only affects processes that are spawned as children of our process,
as the Flatpak process helper deals with this in the session helper.

https://bugzilla.gnome.org/show_bug.cgi?id=773880