runner: add API to set a TTY for IdeRunner
This allows plugins to set a TTY to be used on the IdeRunner. However,
we may need future API for plugins to coordinate on who should be setting
the TTY as the debugger may want to control the TTY in a different view
than the editor.

After you create the PTY slave, simply ide_runner_set_tty() using that
fd to ensure the IdeRunner uses it.