omnibar: Update display name when runtimes are added
Previously, runtimes were loaded strictly before a project was opened so
the IdeOmniBarRow appended "(missing)" to the end of the selected
runtime name if it wasn't found in the IdeConfiguration. But now that
GbpFlatpakRuntimeProvider treats a flatpak manifest file in a project's
source tree as a runtime, we need to be able to update the runtime's
display name whenever it might be discovered asynchronously, so that it
doesn't erroneously appear to be missing to the user. This commit
accomplishes that by connecting to the IdeRuntimeManager's
"items-changed" signal, which is fired whenever the list of runtimes is
modified.

https://bugzilla.gnome.org/show_bug.cgi?id=773764