flatpak: Properly build, install, and run flatpak apps
This commit teaches the flatpak runtime how to use a number of flatpak
commands to build, install, and run applications.

1. In the prebuild worker, "flatpak remote-add" is used to ensure a repo exists
where we can export the build, and either "flatpak-builder" or "flatpak build-init"
is used to build dependencies and initialize the build directory (depending on
whether a manifest was found in the source tree).
2. In the create_launcher function, code was added to use the options and
environment variables specified in the manifest to form a "flatpak build ..."
command.
3. In the postinstall worker, multiple subprocesses are created to run
"flatpak build-finish" to finish the build directory, "flatpak build-export"
to export the build to the repo created earlier, "flatpak uninstall" in case
the app was previously installed, and "flatpak install" to install it on the
host system.
4. Finally, a create_runner function was added that builds a "flatpak run ..."
command to run the app with access to the display even if that wasn't
granted in the build-finish step.

In this way we can exert more control over the build process than if we
let flatpak-builder do all the work for us.

NOTE: Currently this process only works if a manifest is present (so we
know the app id) and at least one git commit has occurred (so
flatpak-builder doesn't complain).

https://bugzilla.gnome.org/show_bug.cgi?id=773764