flatpak: Show prebuild/postbuild output in Builder
In case something goes wrong in one of the prebuild or postbuild steps,
it's useful for the user to see their output in the Build Output window
of Builder. This commit makes that happen by adding an IdeBuildResult
parameter to the runtime's async vfuncs and using it to call
ide_build_result_log_subprocess from the flatpak runtime.

https://bugzilla.gnome.org/show_bug.cgi?id=773764