flatpak: Don't run flatpak-build-init unnecessarily
This commit prevents us from running flatpak-build-init on a build
directory that's already been initialized (which is an error). It also
means we don't run flatpak-builder in that case. An unresolved question
is how to detect if the dependencies have changed (which would mean
the build dir should be deleted and flatpak-builder should be re-run).