flatpak: Use manifest checksum in runtime id
By including a checksum of the contents of the flatpak manifest in its
runtime id, we can ensure that changing the manifset results in using a
different build directory (which allows the changes to take effect).
However this approach has a few disadvantages:
(1) If the manifest never returns to a previous state, that build
directory's disk space is wasted.
(2) The runtime id appears in the omni bar which is ugly.
(3) The runtime id is written to the .buildconfig file which means if
the project is reloaded with an edited manifest, the runtime appears
as missing.

https://bugzilla.gnome.org/show_bug.cgi?id=774509