sourceview: change character overwrite strategy
Previously, we handled overwrites directly in the key-press-event handler.
This was problematic because it is an impedence mismatch with the
GtkIMContext. It was preventing Input Methods from working correctly for
characters like ü (which is Compose→u→").

When fixing things, to filter the event earlier, we were even worse off
because that will swallow almost any thing and commit-text on the context.

Instead, we connect after on the insert-text event and look for a single
character insertion with matching following character. If so, we swallow
it by removing it from the buffer.