flatpak: Delete our remote on application startup
Since the user cache dir is different on the host OS than from within a
flatpak (~/.cache/gnome-builder/... vs ~/.var/app/org.gnome.Builder/...),
the flatpak runtime ends up trying to create a remote with the same name
as an existing one (if Builder has been run outside a flatpak) and a
different location, which fails, eventually causing the Run button not
to work. This commit fixes the issue by deleting the remote we create
upon application startup.