autotools: Check for a dirty config before doing an install
IdeConfiguration has a "dirty" property that we use to keep track of
whether it has changed since the last build so we can, for example, run
the autogen.sh step for an autotools build. But we were only checking if
the configuration is dirty in ide_autotools_builder_build_async but not
ide_autotools_builder_install_async (which also does a build). This
commit fixes that so after the changing the runtime, the Run button should
work.