flatpak: Don't run flatpak-build-finish unnecessarily
It is an error to run flatpak-build-finish on a build directory
muiltiple times, so this commit avoids that behavior. It is possible
that the build affects one of the files that is needed by build-finish,
so we should figure how to detect that situation at some point (and
use a fresh build directory in that case).