langserv-symbol-tree: use node range for tree building
If the Language Server provides accurate information for the range of
the symbol nodes, we can create an n-ary tree structure for the nodes.

This is done by checking if a node contains, in entirety, a child node.
If we discover a node that contains the other node, we rotate the tree
around that replacement.

Currently, rls does not support this due to giving us source ranges that
map to the word/syntax of the document rather than its semantic range. This
is mostly okay because its no worse than things were today where we have
inconsistent "parentName" coming from symbol nodes.