project-tree: Expand tree properly after a refresh
After refreshing the project tree (which checks for new files in the
filesystem), the tree will have newly created IdeTreeNode instances, so
the code we were using to find the node that was selected before the
rebuild wasn't working, meaning that the tree was always collapsed after
a rebuild. This commit changes gb_project_tree_actions_refresh to use
gb_project_tree_reveal, and changes that function to allow expansions of
matching folders, and to behave reasonably for deleted files and for the
the top level folder. It also changes the places where that reveal
function is called to make them work.

https://bugzilla.gnome.org/show_bug.cgi?id=774822