project-tree: Unselect project file after expansion
In gb_project_tree_actions_refresh, we call gb_project_tree_reveal on
the project file to expand the tree, but the project file ends up
selected as well. This commit fixes that by adding an ide_tree_unselect_all
function and makes use of it. (In most cases when you do a reveal it makes
sense to have the node selected, just not this one).