flatpak: add --share=network
This allows network access for building so that things like autogen.sh
can do "git submodule init" which is a common practice. Since we can't
rely on flatpak-builder to do this for the topmost project, we need this.

However, in the future, it would be nice to detect if it is necessary and
only --share=network for the appropriate phase.