flatpak: Refresh list of runtimes when necessary
If the user installs or uninstalls a runtime while Builder is running,
we want to detect that and update the list of runtimes in the Build
Preferences view. This commit accomplishes that using a GFileMonitor.
It does have the side effect that the runtimes are reloaded even if an
unrelated change happens to the flatpak installation (such as installing
or updating apps).

https://bugzilla.gnome.org/show_bug.cgi?id=775863