Loading a project and set it up from a given flatpak manifest
This adds a command line option (-m) to pass a flatpak manifest
when starting GNOME Builder. A greeter with progress feedback
is displayed while the workbench is created and flatpak-builder
downloads the sources in the background.

This works currently for git sources only. Support for archives
and file sources (patches) will be handled in a follow up
commit.