autotools: discover gmake from runtime
If the runtime contains `gmake', we would prefer to use that to ensure we
are communicating with a GNU based make. On some systems, `make' will not
be GNU make but instead might be BSD make.

We cannot rely on this from the configure because our runtime might be
different (such as flatpak runtimes containing only a make and no gmake).