flatpak: Make install work without an Internet connection
Users should be able to build and run flatpak apps locally without an
Internet connection, but this was being prevented because
flatpak-install attempts to check remotes for dependencies. Another
issue is that flatpak prompts the user before installing a needed runtime,
but we're running it in a subprocess. In the future if flatpak gets a way
to install runtimes without user interaction (like -y) we may want to rethink
this and see if we can just use --no-pull when there's no Internet connection.