runtime: add ide_runtime_translate_file()
This allows plugins to get access to a file that is inside of a runtime.
That may be necessary when performing operations such as compilations that
require access to the raw files such as clang. Otherwise, we have no way
to know what /app/include/foo/foo.h means from outside the runtime and
therefore unable to provide valuable code-insight.