flatpak: Detect when the flatpak-builder step fails in a build
Currently, the flatpak plugin assumes that if flatpak-build-init has
been run we can move on past the prebuild step. But if flatpak-builder
runs build-init successfully but then fails when building one of the
modules, we need to be sure to try running it again for the next build.
This commit detects whether flatpak-builder completed succesfully by
creating a file in the build directory; hopefully we can figure out a
more robust solution at some point.