PnlDockOverlay: support popovers in overlay edges
The PnlOverlayDock tracks the focused widget to close
the edge when needed, using a popover in a edge
makes it closing when opening the popover.

The origin of the problem is that we listen to set::focus,
closing the edge on a NULL focused widget and GtkPopover set the
focus to NULL when opening a modal one.