Don't show Other Projects when it is empty
The Other Projects label is shown even if it
is empty when we go to selection or genesis
state and then cancel it to come back to the
browse state.

To fix this we have to call the function
ide_greeter_perspective_apply_filter_all
after changing the state to browse when the
Cancel button is clicked.

https://bugzilla.gnome.org/show_bug.cgi?id=771153