makecache: require makecache file for creation
We no longer want IdeMakecache to be in charge of executing the
make -p -n -s process since that should be done anytime the configuration
changes (and therefore belongs in the build pipeline). This simply allows
us to create a new IdeMakecache from a makecache file on disk. We still
use the runtime to perform various other make commands (like determining
what needs to be rebuilt if a file changes).