autotools: port to IdeBuildPipeline
This removes the vast amount of tricky code we previously had and moves
it into a series of build stages to be attached to the build pipeline.

As the configuration will not change during the lifetime of the pipeline,
all of the work can be performed up front which should simplify some of
our issues with threading in the build process (by making it go away).

We can probably still simplify this further, but this at least gets
everything working.

I would like to see a transient stage added for running `make dist` to
perform tarball creation.