diagnostics: fix the sequence update
The sequence number for a file is imcremented each time
we request diagnostics, that is each time the buffer changes.

Currently, we wait for all diagnostics providers to finish
their task before incrementing the sequence, but we notify
a change after each providers task is finished.

This lead to the following problem:

we can trigger a change without sequence incrementation,
then with the next providers, increment the sequence
but we don't trigger a change.

The result is that the diagnostics in the view are not up to date.

We fix that by incrementing the sequence after each provider's
task is finished.