flatpak: Don't assume the config is a GbpFlatpakConfiguration
If an IdeConfiguration (from buildconfig for example) is using a flatpak
runtime, the pipeline should still be able to do most things. This commit
changes GbpFlatpakPipelineAddin so it doesn't assume the configuration is
a GbpFlatpakConfiguration by getting platform, branch, and sdk from the
runtime instead and skipping parts of the pipeline (like building
dependencies).