flatpak: Use flatpak-build for the runner
We're not doing a full flatpak install to a repo, just building the app
and running the built binary. So this commit fixes the flatpak runtime's
create_runner function to use flatpak-build rather than flatpak-run.
In order to get the name of the binary, it was necessary to add a
get_name function to the IdeBuildTarget interface and implement it for
IdeAutotoolsBuildTarget.