runtime-manager: add ide_runtime_manager_ensure_async()
This function will allow us to check if a runtime is installed or install
it if the runtime could not be found (and a provider supports installing
the runtime).

This is meant to be used by the build manager to ensure that our runtime
is available before setting up the build pipeline.