flatpak: Make flatpak configurations selected by default
If a project has a flatpak manifest, it makes sense to select that as
the default configuration rather than the default Host OS one, since
that's the direction GNOME is headed. In the future, it would be nice to
have a way to prioritize different configuration providers, and a way to
choose between configurations from the same provider.