buildconfig: only use IdeConfigurationManager API from main thread
Most of our libide API is only safe to use from the main-thread. Any
threading we do should be contained to localized objects we control during
the threading lifetime.

This instead keeps a GPtrArray of discovered configurations while we are
threaded, and then defers the registration work back to the main thread
after the threaded task has completed, ensuring we only use that thread
unaware API from the main thread.