progress: make IdeProgress thread-safe
This is useful so that we can use an IdeProgress from a thread (such as
a GTask worker) and have UI updates happen in the Gtk thread. This means
we need to change the ownership of ide_progress_get_message() to always
return a copy.

Additionally, we proxy our property notifications to the main thread.