flatpak: don't add duplicate args to subprocess launcher
gbp_flatpak_subprocess_launcher_spawn() adds any environment variables
that have been set in the launcher to the flatpak-build command line
arguments so they are set in the subprocess. But if the launcher is used
to spawn multiple processes, the --env arguments are duplicated. This
commit fixes that by checking for them in the list before adding them.