makecache: correctly modify subprocess args for 'make -C'
The GbpFlatpakSubprocessLauncher modifies the command line arguments
just before spawning subprocesses, so the position of the subdirectory
(in "... make -C <subdirectory> ...") changes between the first and
second spawn from the same launcher. This causes the
ide_makecache_get_build_targets_worker to replace the wrong argument in
the list when switching to a different subdirectory. This commit fixes
the issue by looking for the "-C" argument each time before replacing
the argument and launching a process.