flatpak: only show recent GNOME runtimes by default
This is GNOME Builder, so obviously we want to be focused on GNOME
runtimes. This defaults to only showing GNOME runtimes for 3.20..master
until the user clicks the "View More" button. Showing all runtimes by
default is absolutely daunting due to all the branches.

If you're building a desktop environment and are using Builder and
flatpak for your *primary and advertised SDK*, then we can look at
showing your runtime as well by default (assuming it requires one).
Presummably you'd also want to register your SDK flatpakrepo with us
as well so SDKs can be installed by default.