transfers: refactory IdeTransfer and IdeTransferManager
These were super difficult to track down issues because of how the
transfer was an interface and all the nastiness that comes with it.

So this changes IdeTransfer to a GObject and drastically simplifies the
transfer manager to remove the unused queuing/max-queue-depth property.

Additionally, there is a signal that is emitted when we want the animation
to occur, so we don't have to use hueristics there. Additionally we can
drop all the "pump row messages/progress" crap now that we can rely on
proper properties from the transfer base class.