subprocess: allow generic FD mapping into IdeSubprocess and IdeRunner
This provides the abstraction we need to be able to map an FD into a given
FD within the target process. Previously we could only do this for
stdin/stdout/stderr.

Additionally, this includes the same plumbing for IdeRunner which is our
more specific, higher-level abstraction for running programs/tests using
subprocess launchers.