valgrind: add basic support for running with valgrind
This is a very minimal valgrind plugin that hooks into the IdeRunner to
run the program with valgrind. The valgrind item should only be available
if valgrind is in your runtime (org.gnome.Sdk ships with valgrind, so it
will be available even when you use org.gnome.Platform).

After running with valgrind, the output will be opened as a document in
the editor containing the results of valgrind.