makecache: fix replacement of -C <directory>
This was pretty complex, for no real good reason anymore. It was probably
due to needing to be relative for when we end up in a different build
directory (such as flatpak), but we map in all those directories now,
so we should be okay with using the proper build directory. Meaning we
can just simply replace the argument after -C.

https://bugzilla.gnome.org/show_bug.cgi?id=771145