pipeline: allow chaining neighboring stages
If we have two stages next to each other during execution that can be
executed as a single step, then we should avoid doing two steps. The one
to think about here is "make all" and then "make install". Not ideal. We
try not to be too aggressive here because we don't want to install if that
phase was not requested.