flatpak: drop bind mount for /run/build/$project
This is where flatpak-builder builds things, but it is terribly
inconvenient for us to do so. It breaks mtimes, it causes make distclean
failures, and is generally just a PITA.

We lose predictable build source locations for debug symbols, but we can
come up with a different strategy for "release quality builds" that get
shipped to users (ie: all flatpak-builder).