build: add distribution channel option for debugging
It really helps us to know how the user got the application, so add a
suffix to the version that gives us a hint to that. If we detect
INSIDE_JHBUILD at build time, we assume this is a jhbuild build. If we
notice .git, but no INSIDE_JHBUILD, we assume this is a local developer
build and they are *NOT* using jhbuild. If it is from a tarball, we should
detect distr (assuming no jhbuild present). And the flatpak builds will
be updated to include --with-channel=flatpak-nightly and such.