IdeDiagnostic: set the location from range when unset
The gnome code assistance plugin creates an IdeDiagnostic object
with no location, and then sets one or more IdeSourceRange's on it.

We can assume that the first source range corresponds to where the
problem is, and set the location based on it. Without a location,
the diagnostic will be dropped on the floor, which is not nice.

https://bugzilla.gnome.org/show_bug.cgi?id=779359