vala: add error regex for vala build errors
This is (what i hope) a someone reasonable regex for matching vala errors
from the console output.

It does expose another error, though, which is that we only get short names
from the valac compiler (main.vala, etc), and we are not translating that
to the proper source directly properly.

That means that when opening the file, we are likely going to get the wrong
path right now ($builddir/main.vala) instead of ($srcdir/main.vala).

https://bugzilla.gnome.org/show_bug.cgi?id=775625