egg-task-cache: Track the in-flight GTask used to fetch the value
Proper cancellation handling, when multiple tasks are queued for a
given key, requires that the in-flight GTask used to fetch the value be
separate from the ones that represent the egg_task_cache_get_async
callers. Fetching of the value should be aborted once all the queued
tasks have been cancelled.

Therefore, it is necessary to be able to access the in-flight GTask for
a given key from a GCancellable::cancelled handler. This is meant to
help with that.

https://bugzilla.gnome.org/show_bug.cgi?id=779660