egg-task-cache: Set up and invoke the in-flight GTask in two steps
Proper cancellation handling, when multiple tasks are queued for a
given key, requires that the caller's GCancellable::cancelled be
monitored. Fetching of the value should be aborted once all the queued
tasks have been cancelled.

It will be nice to cancel the in-flight GTask even before it's invoked,
if the first caller's GCancellable is already cancelled.

Therefore, it will simplify the GCancellable::cancelled handler if the
connection to it is made after the in-flight GTask has been created
and added to the GHashTable, but before it has been invoked. This is
meant to help with that.

https://bugzilla.gnome.org/show_bug.cgi?id=779660