vala: work around Vala language breaks
Commit e874bb7902cc06f9f6d4427d99ec33e3757304e4 broke a bunch of
inheritance in vala (that wasn't very good to begin with, but at least it
worked). In particular, Vala has a very hard time determining if an object
properly meets the interface requirements. It does not handle relaxed
write checks vs the parent very well (and has strange requirements for
vfuncs to be available that are completely unnecessary).

However, we can stop inheriting from Ide.Object and implement the context
property ourselves to work around the issue. Incredibly annoying, but good
enough to fix the build a week before release.