plugins: add GParameter helper
Libpeas is not sufficient to extract GParameters that we need because we
also care about the base objects in the prerequisites, not just interfaces.

This is a sort of barebones implementation of that. Scan the interfaces
chain, and then the objects that are pre-reqs. If we find the property,
all good.

https://bugzilla.gnome.org/show_bug.cgi?id=779891